+++ tests/crq_apis.c
@@ -126,13 +126,13 @@ doit (void)
   s = 0;
   ret = gnutls_x509_crq_get_challenge_password (crq, NULL, &s);
   if (ret != 0 || s != 3)
-    fail ("gnutls_x509_crq_get_challenge_password2 %d/%d\n", ret, s);
+    fail ("gnutls_x509_crq_get_challenge_password2 %d/%d\n", ret, (int)s);
 
   s = 10;
   ret = gnutls_x509_crq_get_challenge_password (crq, smallbuf, &s);
   if (ret != 0 || s != 3 || strcmp (smallbuf, "foo") != 0)
     fail ("gnutls_x509_crq_get_challenge_password3 %d/%d/%s\n",
-	  ret, s, smallbuf);
+	  ret,(int) s, smallbuf);
 
   s = 0;
   ret = gnutls_x509_crq_get_extension_info (crq, 0, NULL, &s, NULL);
